Store data object in iRODS collection

Stores a data object in a iRODS collection. If the data object allready exists, it is not overwritten by default

Usage

ri_put(
  filename,
  collection = ri_getCollection(),
  object = NULL,
  overwrite = FALSE
)

Arguments

filename

filename or path of file to store

collection

name of the collection where to store the data, if not provided it uses the default collection

object

name of the data object, defaults to the basename of the path given by the file argument

overwrite

overwrite existing data objects, default to FALSE