ricmd test function, creates dummy auth file

The ri_test_... functions are meant for package testing and setting up a demo environment. These functions are not intended for package users. creates a dummy auth file

Usage

ri_test_create_auth_file(irods_dir = path.expand("~/.irods"), verbose = TRUE)

Arguments

irods_dir

iRODS configuration dir, defailts ~/.irods

verbose

Be verbose

Due to some quirk in the iRODS python library, an auth file must exist othewrwise a non interactive connection is not possible